+++ src/extraction/features/stratigraphy/sidebar/classes/layer_identifier_sidebar.py
@@ -13,6 +13,7 @@ from utils.file_utils import read_params
 from ...base.sidebar_entry import LayerIdentifierEntry
 from ...interval.a_to_b_interval_extractor import AToBIntervalExtractor
 from ...interval.interval import IntervalBlockGroup
+from ...interval.partitions_and_sublayers import detect_partitions_and_sublayers_with_text
 from .sidebar import Sidebar
 
 logger = logging.getLogger(__name__)
@@ -62,7 +63,7 @@ class LayerIdentifierSidebar(Sidebar[LayerIdentifierEntry]):
         result = []
         for block in blocks:
             interval_block_pairs = AToBIntervalExtractor.from_material_description_lines(block.lines)
-            interval_block_pairs = AToBIntervalExtractor.partitions_and_sublayers_with_text(interval_block_pairs)
+            interval_block_pairs = detect_partitions_and_sublayers_with_text(interval_block_pairs)
 
             result.extend(
                 [
