+++ src/extraction/features/stratigraphy/sidebar/extractor/a_to_b_sidebar_extractor.py
@@ -5,7 +5,6 @@ import re
 from extraction.features.utils.text.textline import TextWord
 
 from ...base.sidebar_entry import DepthColumnEntry
-from ...interval.a_to_b_interval_extractor import AToBIntervalExtractor
 from ...interval.interval import AToBInterval
 from ..classes.a_to_b_sidebar import AToBSidebar
 from ..utils.cluster import Cluster
@@ -60,13 +59,9 @@ class AToBSidebarExtractor:
         pairs = [(entry, find_pair(entry)) for entry in entries]
         intervals = [AToBInterval(first, second) for first, second in pairs if second]
         clusters = Cluster[AToBInterval].create_clusters(intervals, lambda interval: interval.rect)
-        sidebar_segments = [
+        return [
             sidebar_segment
             for cluster in clusters
-            for sidebar_segment in AToBSidebar(cluster.entries).break_on_mismatch()
+            for sidebar_segment in AToBSidebar(cluster.entries).process()
+            if sidebar_segment.is_valid()
         ]
-        sidebar_segments = [
-            AToBSidebar(AToBIntervalExtractor.partitions_and_sublayers(sidebar_segment.entries))
-            for sidebar_segment in sidebar_segments
-        ]
-        return [sidebar_segment for sidebar_segment in sidebar_segments if sidebar_segment.is_valid()]
